backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
alembic
requests
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0